pyproject.toml
[project]
name = "omnisynq"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuito"}
]
license = {text = "GNU GPL 3"}
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "aiofiles (>=25.1.0,<26.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "fastapi (>=0.122.1,<0.123.0)",
    "fastapi-mail (>=1.5.8,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
